# JNIEnv.Initialize optimization