--- a/src/Mono.Android/Android.Runtime/UncaughtExceptionHandler.cs
+++ b/src/Mono.Android/Android.Runtime/UncaughtExceptionHandler.cs
@@ -5,27 +5,9 @@ namespace Android.Runtime {
 
 	sealed class UncaughtExceptionHandler : Java.Lang.Object, Java.Lang.Thread.IUncaughtExceptionHandler {
 
-		static Action<Exception> mono_unhandled_exception;
+		Action<Exception> mono_unhandled_exception;
 
-		static Action<AppDomain, UnhandledExceptionEventArgs>      AppDomain_DoUnhandledException;
-
-		static UncaughtExceptionHandler ()
-		{
-			var mono_UnhandledException = typeof (System.Diagnostics.Debugger)
-				.GetMethod ("Mono_UnhandledException", BindingFlags.NonPublic | BindingFlags.Static);
-			mono_unhandled_exception = (Action<Exception>) Delegate.CreateDelegate (typeof(Action<Exception>), mono_UnhandledException);
-
-			var ad_due = typeof (AppDomain)
-				.GetMethod ("DoUnhandledException",
-					bindingAttr:  BindingFlags.NonPublic | BindingFlags.Instance,
-					binder:       null,
-					types:        new []{typeof (UnhandledExceptionEventArgs)},
-					modifiers:    null);
-			if (ad_due != null) {
-				AppDomain_DoUnhandledException  = (Action<AppDomain, UnhandledExceptionEventArgs>) Delegate.CreateDelegate (
-						typeof (Action<AppDomain, UnhandledExceptionEventArgs>), ad_due);
-			}
-		}
+		Action<AppDomain, UnhandledExceptionEventArgs>      AppDomain_DoUnhandledException;
 
 		Java.Lang.Thread.IUncaughtExceptionHandler defaultHandler;
 
@@ -40,6 +22,12 @@ namespace Android.Runtime {
 
 		public void UncaughtException (Java.Lang.Thread thread, Java.Lang.Throwable ex)
 		{
+			try {
+				Initialize ();
+			} catch (Exception e) {
+				Android.Runtime.AndroidEnvironment.FailFast ($"Unable to initialize UncaughtExceptionHandler. Nested exception caught: {e}");
+			}
+
 			mono_unhandled_exception (ex);
 			if (AppDomain_DoUnhandledException != null) {
 				try {
@@ -54,6 +42,28 @@ namespace Android.Runtime {
 			}
 			if (defaultHandler != null)
 				defaultHandler.UncaughtException (thread, ex);
+		}
+
+		void Initialize ()
+		{
+			if (mono_unhandled_exception == null) {
+				var mono_UnhandledException = typeof (System.Diagnostics.Debugger)
+					.GetMethod ("Mono_UnhandledException", BindingFlags.NonPublic | BindingFlags.Static);
+				mono_unhandled_exception = (Action<Exception>) Delegate.CreateDelegate (typeof(Action<Exception>), mono_UnhandledException);
+			}
+
+			if (AppDomain_DoUnhandledException == null) {
+				var ad_due = typeof (AppDomain)
+					.GetMethod ("DoUnhandledException",
+					            bindingAttr:  BindingFlags.NonPublic | BindingFlags.Instance,
+					            binder:       null,
+					            types:        new []{typeof (UnhandledExceptionEventArgs)},
+					            modifiers:    null);
+				if (ad_due != null) {
+					AppDomain_DoUnhandledException  = (Action<AppDomain, UnhandledExceptionEventArgs>) Delegate.CreateDelegate (
+						typeof (Action<AppDomain, UnhandledExceptionEventArgs>), ad_due);
+				}
+			}
 		}
 	}
 }