--- a/src/monodroid/jni/globals.hh
+++ b/src/monodroid/jni/globals.hh
@@ -3,6 +3,7 @@
 #define __GLOBALS_H
 
 #include "util.hh"
+#include "timing.hh"
 #include "debug.hh"
 #include "embedded-assemblies.hh"
 #include "inmemory-assemblies.hh"
@@ -15,6 +16,8 @@ extern xamarin::android::internal::AndroidSystem androidSystem;
 extern xamarin::android::internal::OSBridge osBridge;
 extern xamarin::android::internal::EmbeddedAssemblies embeddedAssemblies;
 extern xamarin::android::internal::MonodroidRuntime monodroidRuntime;
+extern xamarin::android::Timing *timing;
+
 #ifndef ANDROID
 extern xamarin::android::internal::InMemoryAssemblies inMemoryAssemblies;
 #endif