--- a/src/monodroid/jni/xamarin-app.h
+++ b/src/monodroid/jni/xamarin-app.h
@@ -20,6 +20,8 @@ struct ApplicationConfig
 	bool uses_mono_aot;
 	bool uses_assembly_preload;
 	bool is_a_bundled_app;
+	bool broken_exception_transitions;
+	uint32_t package_naming_policy;
 	uint32_t environment_variable_count;
 	uint32_t system_property_count;
 	const char *android_package_name;