… src/Mono.Android/Android.Runtime/JNIEnv.cs
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Utilities/EnvironmentHelper.cs
… src/monodroid/jni/application_dso_stub.cc
… src/monodroid/jni/monodroid-glue-internal.hh
… src/monodroid/jni/monodroid-glue.cc
… src/monodroid/jni/timing.hh