`JNIEnv.Initialize` is part of the Xamarin.Android runtime startup sequence and
it contributes a sizable chunk of time to it. Some code executed during startup
sequence wasn't as optimized as it could be and this commit aims to pick all the
low-hanging fruit in that area:

  * Don't use the `Logger` class.
    This removes one class instantiation which wasn't absolutely necessary.
    Calls to `Logger` are replaced with calls to a new native function
    `monodroid_log`
  * Use native functions for timing.
    It won't affect the user's code (unless it runs with timing enabled, of
    course) but will allow our timing to be more precise and have less impact on
    code performance. Implemented with two new native functions
    `monodroid_timing_start` and `monodroid_timing_stop`
  * Remove calls to `Environment.GetEnvironmentVariable`
    The calls took a non-trivial amount of time (around 4ms each) and were
    replaced by putting the environment variables in `libxamarin-app.so`, in the
    application environment configuration block.
  * Do not fully initialize `UncaughtExceptionHandler` during init
    The handler's initialization is quite heavy while the handler itself is
    going to be used only when there's an unhandled exception thrown. Delay full
    initialization to the point in time when the handler is actually needed.
  * Avoid parsing of the `PackageNamingPolicy` enum values on runtime.
    Instead, the desired value is placed in the application's environment
    configuration block and simply cast from `int` to the enum.
  * Initialize `Logger` lazily.
    The logging category is requested by `Logger` lazily instead of being set
    from `JNIEnv.Initialize`. This allows the `Logger` class not to be
    instantiated during startup.
  * Avoid static initialization of the `Java.Interop.Runtime` class.
  * Avoid heavy static initialization of the `Java.Interop.TypeManager` class.
    This is done by initializing the type lookup dictionaries lazily, via
    another "proxy" class.
  * Make Java class name lookup faster.
    This is done by moving the lookup (and processing) to native code and
    p/invoking the new `monodroid_TypeManager_get_java_class_name` function.

Other changes:

  * Remove dead code from `JNIEnv`
    The `JAVA_INTEROP` symbol is always "defined" now, remove code used when it
    was not defined.
  * Introduce fast performance measurement for managed code.
    The timing code used by the native runtime is now exposed to the managed
    land, thus allowing us fast and accurate (to the nanosecond level) timing
    measurement.

Impact of the above changes on performance is summarized in the table below.

Device name: **Pixel 3 XL**
Device architecture: **arm64-v8a**
Number of test runs: **10**

|                 | **Runtime init** | **Displayed** | **Notes**                      |
|-----------------|------------------|---------------|--------------------------------|
| **this commit** | 187.907          | 828.90        |  preload enabled; 32-bit build |
| **master**      | 218.804          | 852.70        |                                |
| **this commit** | 154.718          | 803.50        | preload disabled; 32-bit build |
| **master**      | 185.240          | 827.40        |                                |
| **this commit** | 174.211          | 754.50        |  preload enabled; 64-bit build |
| **master**      | 201.779          | 783.20        |                                |
| **this commit** | 144.158          | 741.80        | preload disabled; 64-bit build |
| **master**      | 173.322          | 759.40        |                                |